traslado de la función de extracción  al archivo tasks.py
