+++ apps/entity/utils/general.py
@@ -10,14 +10,6 @@ from django.conf import settings
 # Uso de logger server de django, agrega
 logger = logging.getLogger("django.server")
 
-# Esta dependencia esta solamente en produccion
-try:
-    from sentry_sdk import capture_exception
-except ImportError:
-
-    def capture_exception(err):
-        print(err)
-
 
 def check_exist_act(pk):
     try:
@@ -35,31 +27,6 @@ def open_file(path, type):
         raise StorageFileNotExist()
     else:
         return output
-
-
-def extraer_datos_de_ocurrencias(ocurrencias):
-    # Consideracion: solo se queda con la 1er ocurrencia de estas entiedades
-
-    def buscar(entidad):
-        return next((o.text for o in ocurrencias if o.entity.name == entidad), None)
-
-    entidades = [
-        "CONTEXTO_VIOLENCIA",
-        "CONTEXTO_VIOLENCIA_DE_GÉNERO",
-        "LUGAR_HECHO",
-        "FECHA_HECHO",
-        "EDAD_ACUSADX",
-        "EDAD_VICTIMA",
-    ]
-    contexto_violencia, contexto_violencia_de_genero, lugar, fecha, edad_acusadx, edad_victima = list(
-        map(buscar, entidades)
-    )
-
-    if contexto_violencia or contexto_violencia_de_genero or lugar or fecha or edad_acusadx or edad_victima:
-        try:
-            extraer_datos(contexto_violencia, contexto_violencia_de_genero, lugar, fecha, edad_acusadx, edad_victima)
-        except Exception as error:
-            capture_exception(error)
 
 
 def filter_spans(a_list, b_list):
